requirements.txt
psutil==5.8.0
ray[default]
kafka-python==2.0.2
environs==9.3.2
python-telegram-bot==13.4.1
scipy==1.6.2
numpy==1.19.5